Show delete button only for saved templates

+++ checklist/template/src/main/java/dev/szymonchaber/checkstory/checklist/template/EditTemplateScreen.kt
@@ -375,7 +375,9 @@ fun EditTemplateView(
                             Column {
                                 AddTaskButton(eventCollector)
                                 RemindersSection(template, eventCollector)
-                                DeleteTemplateButton(eventCollector)
+                                if (!ready.isFreshTemplate) {
+                                    DeleteTemplateButton(eventCollector)
+                                }
                             }
                             DropTarget(
                                 modifier = Modifier
